merged `FSHValType` with `DSHType`
All `FSHValType` occurences were replaced with `DSHType`.
Constructors mapping:

FSHnatValType -> DSHnat
FSHFloatValType -> DSHCType
FSHVecValType -> DSHPtr

`DSHPtr` constructor was extended to keep track of
size of memory block (vector) it is pointing to.

In particular, this addresses a shortcoming that
size of global variables was lost during MSH->DSH
compilation.

To make it work we removed avility to pass memory blocks
by value in DHSCOL. This was not an intended feature of the
language anyway.