replace `interp_Mem_equiv_proper` with specialized lemmas
It turnes out that `interp_Mem_equiv_proper` does not hold in general.
There are ITrees which do not preserve equivalence relation on memory.
A possible solution is to replace this lemma with lemmas specialized
for particular ITrees on which it was used and for which we can prove
that they preserve the equivalence relation on memory.